Readme update and code cleanup
